fixed ESP8266 builds

+++ include/plugins/FireworkPlugin.h
@@ -6,10 +6,10 @@ class FireworkPlugin : public Plugin
 {
 private:
   unsigned long previousMillis = 0;
-  const long explosionDelay = 60;
-  const long fadeDelay = 24;
-  const long rocketDelay = 60;
-  const long explosionDuration = 500;
+  const unsigned int explosionDelay = 60;
+  const unsigned int fadeDelay = 24;
+  const unsigned int rocketDelay = 60;
+  const unsigned int explosionDuration = 500;
 
   void drawExplosion(int x, int y, int maxRadius, int brightness);
   void explode(int x, int y);
